fixes, sparc reorg + enhancement
removed:
 lib/sparc64/
 lib/sparc64/atomic.h
renamed:
 lib/sparc64/mempopcnt.c => lib/sparc/mempopcnt.c
 lib/sparc64/popcountst.c => lib/sparc/popcountst.c
	unify the architecture, Sparc v9 is from 1995, since then
	it's 64 bit and has some extras, Sun only keeps everything
	in "compatible till the end of the world" mode
added:
 lib/sparc/memand.c
 lib/sparc/memneg.c
 lib/sparc/memxorcpy.c
	use the Sparc VSI extention (MMX for sparc).
	On older chips it is really a good bit faster, on newer chips it
	shines when stuff is unaligned, the unaligned handling has similarties
	to AlitVec, and actually VSI is the older. Where Intel with the SSE4.1
	palign now finally delivers: unusable crap.

 Makefile
	now cross'ed to ppc64 & arm, which triggered some fixes
 bin2o.c
	make the alignment configureable, some archs have problems with large
	alignments. (which are more tool & format probs...)
 config.h
	typos, doc fixes, give sparc a knob to force v9
 lib/generic/memand.c
 lib/arm/atomic.h
 lib/memcpy.c
 lib/my_bitops.h
 lib/mempopcnt.c
 lib/other.h
 lib/ppc/adler32.c
 lib/ppc/mem_searchrn.c
 lib/ppc/memand.c
 lib/ppc/memneg.c
 lib/ppc/memxorcpy.c
 lib/ppc/ppc_altivec.h
 lib/ppc/strchrnul.c
 lib/ppc/strlen.c
 lib/ppc/tstrlen.c
 lib/sparc/mempopcnt.c
 lib/sparc/popcountst.c
 lib/vsnprintf.c
	typos, fixes, doc fixes
 lib/flsst.c
 lib/atomic.h
 lib/memand.c
 lib/memneg.c
 lib/memxorcpy.c
 lib/module.make
 lib/popcountst.c
 lib/sparc/atomic.h
	unify, add newer arm
 lib/backtrace.c
	"ups, i did it again": arm also has a fucked up greg
 lib/generic/strchrnul.c
 lib/strchr.c
 lib/x86/strchrnul.c
	handle search for '\0' different
 lib/ppc/mempopcnt.c
	opt out of generic code: underlying is a popcount byte instruction.
	If we add up byte sums first, we can save some muls to transform byte
	popcount to over all popcount.

+++ lib/vsnprintf.c
@@ -558,7 +558,7 @@ static inline void print_big(struct big_num *x)
 	i = x->l;
 
 	for(; i >= 0; i--)
-		printf("%016llx", x->d[i]);
+		printf("%016llx", (unsigned long long)x->d[i]);
 }
 
 static int estimate(int n)
